+++ substra/sdk/backends/remote/rest_client.py
@@ -1,6 +1,9 @@
 import json
 import logging
 import time
+from datetime import datetime
+from datetime import timedelta
+from datetime import timezone
 from typing import Dict
 from typing import List
 from typing import Union
@@ -13,6 +16,22 @@ from substra.sdk import utils
 from substra.sdk.backends.remote import request_formatter
 
 logger = logging.getLogger(__name__)
+
+
+def _warn_of_session_expiration(expiration: str) -> None:
+    log_level = logging.INFO
+    try:
+        duration = datetime.fromisoformat(expiration) - datetime.now(timezone.utc)
+        if duration.days:
+            duration_msg = f"{duration.days} days"
+        else:
+            duration_msg = f"{duration.seconds//3600} hours"
+        expires_at = f"in {duration_msg} ({expiration})"
+        if duration < timedelta(hours=4):
+            log_level = logging.WARNING
+    except ValueError:  # Python 3.11 parses more formats than previous versions
+        expires_at = expiration
+    logger.log(log_level, f"Your session will expire {expires_at}")
 
 
 class Client:
@@ -60,10 +79,17 @@ class Client:
             if e.response.status_code in (400, 401):
                 raise exceptions.BadLoginException.from_request_exception(e)
 
+            if (
+                e.response.status_code == 403
+                and getattr(e.response, "substra_identifier", None) == "implicit_login_disabled"
+            ):
+                raise exceptions.UsernamePasswordLoginDisabledException.from_request_exception(e)
+
             raise exceptions.HTTPError.from_request_exception(e)
 
         try:
-            token = r.json()["token"]
+            rj = r.json()
+            token = rj["token"]
         except json.decoder.JSONDecodeError:
             # sometimes requests seem to be fine, but the json is not being found
             # this might be if the url seems to be correct (in the syntax)
@@ -71,6 +97,13 @@ class Client:
             raise exceptions.BadConfiguration(
                 "Unable to get token from json response. " f"Make sure that given url: {self._base_url} is correct"
             )
+
+        logger.info(
+            "Logging in with username/password is discouraged; "
+            "you should generate a token on the web UI instead: "
+            "https://docs.substra.org/en/stable/documentation/api_tokens_generation.html"
+        )
+        _warn_of_session_expiration(rj["expires_at"])
         self._headers["Authorization"] = f"Token {token}"
 
         return token